Updated openstack/openstack Project: openstack/tripleo-incubator  2c401ea50a54c62a76cf5580c09fec6ee6ab43c6
setup-network: add support for vlan trunks

This patch adds a new --vlan-trunk-ids option to the
setup-network script so that can be used to enable trunked
VLAN IDs.

The goal is to allow VLAN traffic to flow between our VMs
in a virtualized development environment.

Normal untagged traffic is still allowed on VLAN ID 0. This
should allow normal ctlplane provisioning traffic to continue
flowing if VLANs are used on the same bridge.

Change-Id: Ic5fbb2a54049c6e6bd7a64317268cf79b21020ba