Updated openstack/openstack Project: openstack/governance  7d5be035c542bc137c0b227e0f2da33a16b44061
Move application to current projects up in tag page

Feedback from operators in a user committee session pointed out which
projects a tag is applied to is one of the most useful things on each
tag's page.  So move the 'application to current projects' section up in
the page so it is easier to find.

Change-Id: Ib7b0aa250f950398e0abd20d35a1ea252b057d4c