Updated openstack/openstack Project: openstack/governance  c32f747053ad0cef4d24e0282602f5e15c91f538
Add networking-bgpvpn repo to Neutron.

The networking-bgpvpn repository, currently in stackforge, is a Neutron service plugin
for interconnecting Neutron networks and BGP-based VPNs.

https://github.com/stackforge/networking-bgpvpn

Following discussions during the summit, the identified next-step is to move this proposal
under the Neutron "big tent" (related docs about adding repositories like this to Neutron
can be found here: https://review.openstack.org/175952 ).

The implementation is an early implementation of specifications initially proposed for
the Neutron main repo for Liberty release.  The specs will move in this new project and
the implementation will be updated to reflect comments made on the spec:

https://review.openstack.org/#/c/177740/

This implementation is designed to delegate the actual implementation work to
drivers, targeting for instance such or such "SDN controller". A lightweight
implementation based on bagpipe-bgp [1] is already provided and suitable for running
CI tests.

Change-Id: I6817d63e489409edf382eba2f947f3451d9bd5fe