Updated openstack/openstack Project: openstack/ironic  f87bad418f88a6a25c01c37285f9b8af3ff9e1ab
Remove bogus conditional from node_update

We have a conditional in the node_update() method checking if a user
is trying to modify the power_state of a node via PATCH, that won't
never get triggered because power_state is an internal attribute and
the request will fail on the API side.

Plus it makes no sense for checking power_state but not other states
such as provision_state, console_enabled, etc...

The NodeInWrongPowerState exception was also removed because it's not
being used anywhere else.

Change-Id: Ibc726fb9b91af8ca0e3880350d2a499b9501a7f3