Updated openstack/openstack Project: openstack/tripleo-heat-templates  4fe67fdc38f66f2ee03423b8154dbb19baced46a
os-net-config templates to configure vlans

This patch adds 5 new role templates to help configure
a vlans on top for each of the overcloud roles. This
patch adds vlans on top of a single NIC attached to
the control plane network (already used for provisioning).

The patch also includes an environment file to
enable configuration of vlans by simply sourcing this file.

Change-Id: Ibc40e452dec9b372ff10442aab2bddaf382b0a2f