Updated openstack/openstack Project: openstack/sahara-specs  a10c5176e1da6ccb8c439ef64608c96462ec6266
Removing JobExecutionArgument Table

On implementation of the Unified Job Interface Map story, the
JobExecutionArguments table proved fairly useless. As the
JobExecution object is frozen by the time it gets to the edp
engine, the job_configs json must be updated, and thus no real
record-keeping purpose is served. Removing it for simplicity.

Change-Id: I010290e8c52b016dfa2f4d7563038bd7407563e5
Describes: blueprint unified-job-interface-map