Updated openstack/openstack Project: openstack/neutron  2fc39b3531fa4822f7590d9b6a5f6d8da619e7a4
Make MockFixedIntervalLoopingCall class as a helper class

MockFixedIntervalLoopingCall class is used in both the classes
TestOvsNeutronAgent and TestOvsDvrNeutronAgent.

This patch removes the MockFixedIntervalLoopingCall from both the
places and make it as a helper class in test_ovs_neutron_agent.py.
So that it will be used everywhere in test_ovs_neutron_agent.py.

Change-Id: I5e0803436e91b4f20ec03209e9e799a68df4eaf4