Updated openstack/openstack Project: openstack/python-openstackclient  1fff11a9ab409a62fd48ee235377deb4c400c59f
Add EC2 support for identity v3 API

EC2 support is provided for the v2 identity API and is available in
almost exactly the same format in the v3 API and enabled by default.

Supporting EC2 in the v3 identity API in OSC will make it much easier to
transition devstack to a v3 only state.

Closes-Bug: 1236326

Change-Id: I52ff2020ef2fcbdc8a98280b73c6fd4a93bc8e0f