Updated openstack/openstack Project: openstack/requirements  b0925cc208bfd0a3ca5e1e858adbfa5a24b96ac7
Consolidating redundant code in the tests

This moves things that are pretty much copy-paste into one place,
letting us focus on the actual differences amongst the tests.

I've also made the local test fixtures explicit, so we can see what
each test really needs, which has also eliminated redundant test
overhead, the slowest test is now 0.05s vs 0.1 before. This is more
than paid for by the reduced duplication in the basic setup, making
the patch a net reduction in LOC.

It also fixes the tempdir leak that was present in two of the test
scripts.

Change-Id: I63eee7681657b2b5eb7959738cc89c57c117fa3b