Updated openstack/openstack Project: openstack/keystone  a637ebcbc4a92687d3e80a50cbe88df3b13c79e6
Don't query db if criteria longer than col length

When a long query parameter name is used (such as a name with
more than 255 characters), this is never going to match a user
in the SQL backend, because names can only be up to 255
characters. Keystone doesn't have to bother querying the backend
in these cases and doing so is unnecessary. With this change,
these types of requests are short-circuited.

Some databases (e.g., DB2) return an error if the query string is
longer than the field.

Change-Id: I68b3abaabec6cb46a9a320e312c0be7707009170
Closes-Bug: #1386773