Updated openstack/openstack Project: openstack/tripleo-heat-templates  5c315de4bccd397b0a8b704bb2a48c62640e2370
Wire ServiceNetMap as a top level parameter

This patch makes ServiceNetMap a top level parameter.

This is helpful to tools like Tuskar which don't support Heat
environments that contain both a resource_registry and default_parameters.

ServiceNetMap will in fact be utilized at the top level in some of
the VIP related patches that follow.

Change-Id: I375063dacf5f3fc68e6df93e11c3e88f48aa3c3a