Updated openstack/openstack Project: openstack/tempest-lib  3fe26726c099c488e2ef7d6dd8ca60881fca8c4c
Switch to using ostestr for subunit-trace

The subunit-trace module has been moved into the os-testr package.
This decouples the nominally independent module from the rest of
tempest-lib which should enable wider usage. However, since some
packages are depending on tempest-lib to use os-testr at this point
we add it to the requirements.txt file instead of test-requirements
so that those consumers won't be broken when this is release.

Additionally, os-testr includes a testr wrapper, ostestr, which
replaces the pretty_tox.sh script and is used to run unit tests.

Depends-On: Ia399c65e62c2306c9a2d082c6219fcb030b5946b
Change-Id: Iafbfa92a91b945b75aaa1f684594c903c8f6f700