Updated openstack/openstack Project: openstack/horizon  50d80bb41307c3d0b7c130c0d76fd2d59d268ad2
Fixing up how job configs are handled

There are a handful of job configs that are handled by
other controls on the config form.  We now explicitly
prevent them from showing up in the free entry inputs
while allowing all the unhandled configs to appear when
attempting a relaunch of a job.

Change-Id: I493f21500f76bc296c5d41b2b624be335243414f
Closes-Bug: #1445228