Updated openstack/openstack Project: openstack/keystonemiddleware  6f10b9c8e1dbcd3450b33cae6126961d60e199d0
Base use webob

Why are we manually managing environment vs headers? We already have a
dependency on webob, let's actually use it and simplify our code a
little.

This is a very basic patch which just allows us to use webob requests
and responses to prove there is no difference. From this we can do
additional cleanups in later reviews.

Change-Id: I35c773fb4a15a1c8391c4129300ef2f0592d79ee