Updated openstack/openstack Project: openstack/tempest-lib  86620669b329ed75378b2b78f6b6af7378458888
Add unit tests for basic auth methods in token clients

This commit adds the first missing unit tests to the token clients.
Previously there was no coverage at all for these clients so
there was no in-tree method for verifying the functionality of these
classes. The coverage of this commit is limited and only covers 2
basic uses of the auth() method in both the v2 and v3 token clients.
It will be expanded in future patches. As part of enabling this
coverage the fake_http class had to be modified to return a user
provided body and response code.

Change-Id: I1888b31550e5db4343003b05c0f6865444756789