Updated openstack/openstack Project: openstack/horizon  9d12f26d0b3d70a9442f473df150c401f24177ce
ngReorg - move utils from dashboad-app

This will facilitate the removal of dashboard-app from
horizon-the-framework.

The "tech-debt" folder is intended to be removed in subsequent
patches (still to be created) which move or remove the functions
in that folder. The reason for not doing so now is that there
would be functionality changes, and I wanted to keep this patch
as small as possible.

The dummy controller also moves to the same place as it
is an empty controller that will eventually go away when
the Django modal it's used in goes away.

Change-Id: I0b89205b9c6c7b1498c1a3c48e53ba460b9e5469
Partial-Bug: #1458697