Updated openstack/openstack Project: openstack/keystone  b19171eca844f0d9b2ab9bd9592b3259575fd388
Order routes so most frequent requests are first

Since authentication is going to be the most common operation in a real
environment, I moved this first.

The problem is that routes uses one, yes one, huge regex for
dispatching. This is not performant for the unfortunate routes that
appear at the end.

Change-Id: I06bc2fa6a29934e28ab83d92bc9bf9b3665ba34c