Updated openstack/openstack Project: openstack/nova-specs  e32744eeac8011bdc1adcaf0ff629fa214876e4d
Use os-brick library

This spec lays out the process for which we remove the internals
of the libvirt volume drivers to use the new os-brick library that
has been incubated inside of Cinder since the Havana release cycle.

Cinder has been updated to use the os-brick library already.

Change-Id: Ia5ea709dc4749e207b0278e5d5096ba0154a31d8
Blueprint: use-os-brick-library