Updated openstack/openstack Project: openstack/nova  8f04566e456c618971512d9566a303541e10d30c
Raise InstanceNotFound when save FK constraint fails

Currently, in the Instance.save method, if an ObjectField is dirty,
it will be updated in the db before the instance record is updated
(separate transaction). If the instance does not have a row, the
foreign key constraint fails and DBReferenceError is raised from
oslo.db. This scenario can happen in cells if an instance delete
request comes in before it's scheduled, during the broadcast to
delete in all cells. This is expected and InstanceNotFound is
caught and handled.

This change catches DBReferenceError and raises InstanceNotFound.

Closes-Bug: #1462128

Change-Id: I821f8c121760e0fb74d7ae5c680bd0c361f92128