Updated openstack/openstack Project: openstack/heat-templates  2636f415b85e4689f832541cc79b4cd00711fbaf
Improve handling of Puppet output

Currently the output of Puppet run is written to os-collect-config log only.
If the run gets stuck, we have zero output from it, and have no chance
of guessing where it got stuck. This patch is fixing this issue and also adds
support for running Puppet in debug mode.

Due to big amount of data stdout and stderr in response during debug mode,
both are trimmed so that we won't exceed  max_json_body_size(1048576).

Resolves-bug: rhbz#1242396

Change-Id: Ie92d1714a8d7e59d347474039be999bd3a2b542f