Updated openstack/openstack Project: openstack-infra/project-config  15aad4b870cdc53b974731797039cc7e28f52910
Add #openstack-neutron-release channel

Since Neutron is implementing it's own verison of the "Big Tent"
governance model (called the "Neutron Stadium"), I'd like to create
a new IRC channel we can use to sync release activities and to use
as a sync point between the projects. This will also be used for
syncing between PTL and Lieutenants during an office hours window.

Change-Id: I22748294669c6d402751448597c823f4bed66bbe
Signed-off-by: Kyle Mestery <mestery@mestery.com>