Updated openstack/openstack Project: openstack/cinder  5c2ebb02e57a481050b7e44453b58e8a306536ac
Windows SMBFS: Fix image resize errors during volume creation

When creating volumes from images or snapshots, the corresponding
virtual disk image is resized up to the volume size.

At the moment, the resize is performed even if the image already
has the requested size. This can raise an excepion in some
environments.

This patch checks the image size and performs the resize only if
it's required.

Change-Id: Ib65f0e8d69f93ae21975eb46ba24e085c06e4e34
Closes-Bug: #1437306