Updated openstack/openstack Project: openstack/horizon  54ee33c4d844961f7f7f7d94b3c298a015b09e95
Handle modals properly when disabling table actions

This fixes a problem when attempting to disable table actions on
a modal dialog. It causes the Launch Instance modal to render
incorrectly and other modals result in a
"$form.on is not a function" error displayed in the browser
console. This is because the modal init function is assuming the
argument passed in is a jQuery reference to the modal but in
reality it is just the modal DOM node itself. The problem is fixed
by making a jQuery reference to the modal before preparing to
disable the actions of any tables that might be displayed on it.

Closes-Bug: #1462544
Change-Id: I56399468ff5e2b5bcd0b285ece0d6dd001500c24