Updated openstack/openstack Project: openstack/os-collect-config  b63188eb25f5289ad6526c77a7c4fe8ffcd022a1
shorter sleeps if metadata changes are detected

This patch updates os-collect-config so that the sleep interval
time is shortened if changes are detected. This should decrease
deployment time when using Heat templates which use depends_on
to step through a sequence of software deployment resources.

The new default sleep is set to 1 and increases
by sleep_time *= 2 until it reaches the default
sleep interval again.

Change-Id: I5cbd0956db2abebec876b15bee72b70ec64d5aef