Updated openstack/openstack Project: openstack/horizon  2eeba491c4c636400acc5a255fd7dd9d91090b32
Refactor quota limit summary template

Refactor the quota limit summary template to allow easier customisation.
Also change it so that it removes the 'of' when there is no quota limits in
place so that it just reads Used 'x' (No Limit).

Added tests for checking usage['charts'] is in the context and that (No Limit)
is present when there is an infinite quota on the page.

Change-Id: Iba62ca89e53f0d6b9c583e4cc450e853c18ba107
Closes-bug: 1439717