Updated openstack/openstack Project: openstack-infra/project-config  d6d21e35ad5ebf6af8c37b66ca566191dab8e7c1
Add libuuid devel packages

python-qpid-proton does something similar to pyzmq when libzmq is not
installed. It bundles the underlying library and builds it.
Unfortunately, python-qpid-proton currently has a dependency on
`libuuid` and therefore, it depends on the devel packages to be present.

This dependency will be removed in the future but it'll take time. Until
that happens, I'm adding uuid-dev to the list of packages to install by
default.

See related system-config change: Iae4d7138311eec0cd90c0865f30b84a98ce3c142

Change-Id: I65f24262c2470c6d75b5b41106469604f5b105c4