Updated openstack/openstack Project: openstack-infra/shade  1bc6f5d351daaea1eaeaefa120f07bae50205ad7
Extract logging config into a helper function

It's an error to have a library class set up logging output for you, as
this will prevent a consumer from doing what they need to do with their
logging config.

However, for simple scripts, the complexity of what needs to be done for
"normal" operation is a bit onerous.

Create a simple helper function that a user can run to set up the two
most common types of logging expected for simple scripts.

Change-Id: I81ef597712ff885d95194c6e29a45a1b2e7f86b9