Updated openstack/openstack Project: openstack/nova  770a897179f09546b32763697319d49be1697be1
Begin the transition to an explicit object registry

In oslo.versionedobjects, the registry is opt-in instead of implicit.
This patch sets the stage for moving to that by defining a no-op
registry with a compatible register method so that we can go ahead and
decorate the object classes and then move the registry underneath them.

Related to blueprint use-oslo-objects

Change-Id: Iddc68a2d8a9b103621bc60a7426fe495edf80aad