Updated openstack/openstack Project: openstack/heat  f8aabd3c18736f7f7b2c7db506d74ac1c36219df
Move manila resources in-tree

This change relocates the manila resources from the contrib area into the main
resource tree. It was originally added to contrib/ because of the project's
incubation status, and more specifically because the client is not in the
global-requirements.txt file. However, when this was discussed at the summit in
Vancouver, the decision was to move the resources to main tree but skip
registration if the client is not installed. This will save users from the
trouble of installing it as a plugin.

Change-Id: I9f43eba2abea351e1c5427b0968cbccab355992c