Updated openstack/openstack Project: openstack/nova  4630310974b908bb303da426547e40ae66dc8b27
Merge fping functional tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges fping functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled tests to run
merged functional tests for both v2 and v2.1.

Change-Id: I1dd91402c38736accea0e9b73be66a9aaa72c4f3