Updated openstack/openstack Project: openstack/nova  084b8c79f2be1f5f3e242ceadb1ceb1537cc0ff9
Share admin password func test between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges admin password functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled for v2.1 tests to run
successfully.

Change-Id: Icfb9253ffb0852d15e685045a31b9fe811722733