Updated openstack/openstack Project: openstack/tempest-lib  f46b6c99e05c2a4dba23f071edc2b1dad23e42dc
Fix 'format' checking for json schema validation

In json schema, "format" keyword is defined to allow inter-operable
semantic validation for a fixed subset of values.
For example - ip address which comes as 'string' can be further checked
with ipv4 or ipv6 format.

ip_address = {
    'type': 'string',
    'oneOf': [
        {'format': 'ipv4'},
        {'format': 'ipv6'}
    ]
}

Json schema does not validate 'format' correctly if FormatChecker is not
passed during validation. Without FormatChecker, it validate all format as true.

To validate 'format' correctly FormatChecker has to be passed while doing schema
validation.

This commit fix this issue and adds unit tests for that.

Closes-Bug:1460975

Change-Id: Iad6ec9644eb1ec976b4a47f7dec127bfaecfd1fa