Updated openstack/openstack Project: openstack/nova  357b5660151d26578df7a54c57f4b59e038b06c9
Fix an objects layering violation in compute/api

This patch fixes a case in compute/api that "knows" the magic name for
where we store actual object contents.

I'm not sure about this actual use and why we need it, but in order to
not block subsequent patches, this makes it at least use the proper
API in order to generate the magic attribute name.

Related to blueprint use-oslo-objects

Change-Id: I4fa95f03106130c64b672c32247686d109c760f7