Updated openstack/openstack Project: openstack/nova  0635c4a1d2ec5e0d9157115907ec4fb139d03a48
VMware: remove unused parameter for VMOPS spawn

There is no longer a need for the power_on paremeter to be passed
tp the spawn opertaion. This was originally used for the rescue but
commit cd1765459a24e52e1b933c8e05517fed75ac9d41 ensured that we no
longer need this.

This patch remove a parameter that is no longer needed and cleans
up tests related to this.

TrivilaFix

Change-Id: If719b94cf50d5f2212eac4807c80edb886bf7808