Updated openstack/openstack Project: openstack/ceilometer  32936ae138ec113662d51d13324efe83b7951102
Cause gabbi to skip on no storage sooner

This avoids problems interacting with the global CONF used by
prepare_service that happen when the fixture is skipped. If we skip
before ever accessign the CONF we have a better known state.

This problem was discovered by comparing runs of the gabbi tests
with mongo and mysql (they only work with mongo, by design).

Change-Id: I9567fb0b37e68cd92f141697e4af8882c10a33f7