Updated openstack/openstack Project: openstack/keystone  72ac6259965f7a455413cb971ab5ae518d594900
Add testcases to test DefaultDomain

As to the external authentication methods, we have `KerberosDomain`, `Domain`
and `DefaultDomain` defined, `DefaultDomain` is a different usecase compared
with `KerberosDomain` and `Domain`.

We have testcases for `KerberosDomain` and `Domain`, but there is no testcase
to verify `DefaultDomain`, so we need to fix it up.

Change-Id: Iec3e9a44c0b6fa22305aaf05f2f2aa9bffbf7b49
Closes-Bug: #1459116