Updated openstack/openstack Project: openstack/nova  2aad224d2c4f9d8b5aa37e2f09e664a017411b8c
docs: fixup libvirt NUMA testing docs to match reality

The libvirt NUMA testing docs were written before the final code
was merged, so the example JSON shown for the database fields
did not match reality. It was also hard to read the JSON since
there were no line breaks & indentation. To make matters even
worse, the final example was just plain wrong - the data shown
was a mixture of a guest with 1 NUMA node and a guest with 2
NUMA nodes.

Change-Id: I51e4df2d9fe252ba65f6a37d84f125e18b44e239