Updated openstack/openstack Project: openstack/requirements  804a6b42f36d099540437213c0e9fa7a8a5918da
Raise minimums for oslo libs without namespace packages

Update the minimum versions for several oslo libraries that no longer
use the namespace package.

The oslo.middleware version that dropped the namespace package is blocked
because it causes issues with our upgrade when the paste.ini files are
not upgraded when/before the code is upgraded.

Change-Id: I7ff3e9aaf7196efd8814e068b5dc52d1dd2ffe04
Blueprint: remove-namespace-packages