Updated openstack/openstack Project: openstack/nova  143afe26ddae74b0b56573cb9bcf071513a6527d
compute: only use non_inheritable_image_properties if snapshotting

The CONF.non_inheritable_image_properties parameter lists a set of
image metadata properties that should be discarded when uploading
an image to glance during the snapshotting process.

This property is currently applied in the nova.utils method
get_image_from_system_metadata(), which causes the properties in
question to be discarded in a wide variety of non-snapshotting
related operations.

This is a regression caused by two commits:

  commit 8e575be75c80ea71a6ad8fb73e6ace1ed708938f
  Author: Xavier Queralt <xqueralt@redhat.com>
  Date: Mon Aug 26 22:53:03 2013 +0200

    Add methods to get image metadata from instance

  commit 4389f2292a0177c8eedc0a398ceb3c5535a9ef82
  Author: Xavier Queralt <xqueralt@redhat.com>
  Date: Mon Aug 26 22:55:46 2013 +0200

    Avoid errors on some actions when image not usable

Which moved handling of CONF.non_inheritable_image_properties out
of the compute API _create_image() method, and into nova.utils.

Fix this by moving the handing back into the compute API in its
original location.

Closes-bug: #1459760
Change-Id: Id630fe68678c4aa469ddbfdb3c757b264519e918