Updated openstack/openstack Project: openstack/nova  0557da06911811c1e647b22653a915d31b45b2b7
objects: fix parsing of NUMA cpu/mem properties

When processing hw_numa_cpus.NN/hw_numa_mem.NN properties, the
ImageMetaProps object iterates from 0 through to the value of
hw_numa_nodes. If any NN is missing, it leaves it uninitialized.

This is wrong on two counts, first the hw_numa_nodes value is
not required to be set in the image - it can come from the
flavor, so we must not rely on it existing. Second, setting
of sparse topologies is not supported - all nodes must be
specified.

Change to iterate from 0 through to NUMA_NODES_MAX, defaulting
to 128 for now since that ought to be large enough for a good
few 10's of years to come. Stop reading at the first node for
which no data is present.

Change-Id: I2155164ad59a9b68245071d48d9dd501b76f7499