Updated openstack/openstack Project: openstack/ironic  070cc9ecbf8e474becb676f8d8180d9d3812225a
Merge tag '2014.2'

This is a null-merge of the 2014.2 release tag back into the
master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Icf942802a151d4816ce38fac343a70375720ed47