Updated openstack/openstack Project: openstack/trove  3c8112735d0e280efbc6e23ac5f018c02bdd48d2
correct api schema for instance patch

The api schema specification for the patch oepration is in error and
defined a parameter of "configuration_id" while the code (both in the
server and the client, and also all the tests) used "configuration".

A fix with fewer lines of code would have been to just embed the {
"type": "null" } into the definition of configuration_id but I am
going to be making additional changes to the way in which we signal
(through a put call) that we want to detach a configuration_id so I've
defined a null_configuration_id type which I can use there as
well.

To prevent this from reoccuring, additionalProperties has been set to
False.

Change-Id: Ic98050e487b0d8cd9a68e8c46169456a52e2b16c
Closes-Bug: #1460174