Updated openstack/openstack Project: openstack/glance  74aeb021c44021722825d1f2476f4756aeff69ff
Mark task as failed in case of flow failure

When a taskflow flow fails at an intermediate task before complete
task, the glance task is in the processing state indefinitely. This
patch marks the task as failed.

In addition, it also fixes the importToFS task to properly analyze
the taskflow result object.

Closes-Bug: #1459015

Change-Id: Icd325c87bac5a48ea177cfa883a173c3c490feb2