Updated openstack/openstack Project: openstack/governance  09a028a8f44d52f25db21d6e073de2808870a07d
Add the Searchlight Project:

Searchlight is a new project being split out of Glance based on the
Glance Catalog Index Service, which was developed and released in
Kilo [1]. We received community and operator feedback that this
would be very useful for more than just Glance. At the Liberty
Summit it was  decided to broaden the scope and make it its own
project with a mission to provide advanced and scalable search across
multi-tenant cloud resources. This was presented and discussed
at both Glance and Horizon fishbowl sessions dedicated to this
topic where it was enthusiastically received.

A narrated screencast of the demo shown at the summit is
available at the project wiki [2].

The project is being launched based on a deep history of following
all of the principles of the OpenStack way:

While in Glance, 21 Engineers at 6 different companies contributed in
design sessions, code, or related reviews to the Catalog Index
Service. The concepts that directly led to it becoming a part of
Glance were first presented at the Juno summit, taking some
inspiration from search concepts previously discussed at the
IceHouse summit.

All code has been developed under an Apache license and all
library dependencies allow for unrestricted distribution and
deployment. It has been developed in Gerrit under the Glance project
with only this initial separation from Glance being copied to github.
No further work will be done on the codebase until the project is
back under the Gerrit system.

All previous IRC meetings have been held in the Glance IRC meetings,
but now have an approved IRC meeting [2] of its own in
started using #openstack-searchlight for open discussion.

The project will initially have a co-PTL with Travis Tripp and
Nikhil Komawar (current Glance PTL) to launch the project and regular
elections will be held to elect the PTL. The initial set of drivers
(currently 12) is based on shared interest and contribution to the
service while it was in Glance.

Searchlight is compatible with OpenStack APIs, and uses Keystone
middleware to integrate with OpenStack Identity. It also consumes
notifications from other OpenStack services and implements a plugin
mechanism for resource abstraction.

[1] http://specs.openstack.org/openstack/glance-specs
    /specs/kilo/catalog-index-service.htm

[2] https://wiki.openstack.org/wiki/Searchlight

[3] https://review.openstack.org/#/c/186060/

Change-Id: I8b44aac03585c651ef8d5e94624f64a0ed2d10b2