Updated openstack/openstack Project: openstack/oslotest  ecad0651b1c9a269785d645d839bbd6113632cf7
Document the mock attribute for mockpatch

When using mockpatch.Patch/PatchObject you need access to the
mock in order to verify that it's getting called (e.g.,
assert_called_once_with), but this attribute wasn't documented so
nobody can tell how to do it.

While adding the attribute, other issues with the docs are
corrected.

Change-Id: I7df083998adcee710cf1c5b87a1b8f091e7ad522