Updated openstack/openstack Project: openstack/horizon  084ab1b033fd6c771b684a333d1dc3cf5df15b5c
Escape the description param from heat template

The heat template allows user to define custom parameters,
the fields are then converted to input fields. The description
param maps to the help_text attribute of the field.

Since the value comes from the user, the value must be escaped
before rendering.

Co-Authored-By: Lin Hua Cheng <os.lcheng@gmail.com>
Change-Id: I79d540a8363b2507c4bccdc0cc38e283962919d2
Closes-bug: #1453074