Updated openstack/openstack Project: openstack/barbican  146e0350f797248c31156a5473fc63450ab41752
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iba4587072d48836ee0d5cf7bf9b176bd0c2d6545