Updated openstack/openstack Project: openstack/ironic  6ebcd05e03b3d232616ec236a2651cb66700e99a
Fix DRAC driver job completion detection

The DRAC driver code determined job completion by comparing the
returned JobStatus to "completed" and "failed".  Testing with hardware
showed that JobStatus can also return "Completed with Errors".

This change updates the DRAC driver so that it recognizes a job with
"Completed with Errors" for the JobStatus as a completed job.

Change-Id: I051d6703194231712f0234f31b0b49208bf0e6ec
Closes-Bug: 1462513